Add files via upload
These are sample files of GPS data collected from the threads reading TCP/IP ports.  

Sample2output1.txt is the thread assigned to GPS1
Sample2output2.txt is the thread assigned to GPS2

This data is typically coming over TCP/IP into the program.  What is lost in the capture of the file is the timing and network impact of and jitter or delays.  It is critical that the timestamps of both GPS's be aligned in the reading file and that the reading and calculations are done with minimal delay to output the new NMEA sentence as quickly as possible.